Overview - don't update task overlay on thumbnail update.
Only init and reset the task overlay on the enabled state of the overlay
changing.
Previously the overlay was also updated when the thumbnail changed, but
this can happen for lots of different reasons so was causing janky
transitions and potentially excess work in code that was following the
overlay.

The overlay code really only cares about the thumbnail after the task has
settled, so should be safe to ignore the thumbnail updates.



Bug: 188097940
Test: local build and run on p5
Change-Id: I4ad1767046cb0316af19713bc188788d8f581d7c