Register a remote animation for the next activity start
Widgets are permitted to start broadcast receivers which launch
activities, otherwise known as a broadcast trampoline. Unfortunately,
launch animations provided through ActivityOptions are not applied to
the activity launch.

If a widget interaction's pending intent is not an activity, set a
return animation for the next activity start, in the event it starts an
activity.

Bug: 169042867
Test: presubmit
Change-Id: Ibbee9cadcd5153e285433fcb1b8038d77b1384a2