Force initialize task stack callback for gesture modes on init
This handles an edge case where RotationTouchHelper#mMode starts as 3 button (by default)
* then changes to gesture nav.
* then onDestroy called
* then on re-initialization mMode will be gesture nav instead of 3-button
(it maintains state because of it being singleton).

The check to register the callback won't happen because we register only if previous mode and
current mode differ.
Don't rely on default value of mMode being 3 button when initializing

Bug: 181928969
Test: Followed repo steps in bug.
Restarted phone in all 3 nav modes and verified
expected quickswitch behavior

Change-Id: I381eba7417c4cde593edf4b9a34da55b50774e98