Fix foreground position during widget home animation
The home return animation applies a position animation to all views in
the workspace. Since the widget uses a GhostView of the original widget
to represent the foreground of the animation, that animation was
reflected in the foreground's positioning.

The CL creates an ignore list of workspace views to which an animator
should not be applied.

Bug: 190488979
Test: manual
Change-Id: I5a86165e945336693eedd253a02bda1e1eb6662d