Send a zero for assist invocation on animation end
When a corner gesture doesn't complete, we animate the progress from its
current value back to 0. When it hits 0, sysui will perform some cleanup
e.g. removing the window. However, it's possible the animator won't
return exactly 0 and will complete, so we should send an explicit 0
whenever the animator completes.

Bug: 187768368
Test: We don't have a repro for the bug, but expect this should be more
      robust. Tested to validate it didn't regress.
Change-Id: Id65f94534363c326fc2c8d0a9114a3c036aa55e4