[Live Tile] App surface is above launcher even after reaching Overview
Currently we place app surface below launcher and punch a hole through task view to display live tile in overview. In this change the order will be flipped, with a few benefits not limited to the following. For example, we have a surface view lifecycle bug where the surface view in a live tile app can go blank during app launch transition. This change will significantly mitigate the issue visually as the background wallpaper will not be revealed.

In order to make sure visually everything still looks the same, we will need to finish the recents animation when the user taps on the task view icon, since the menu overlaps with the task snapshort view.

This change also addresses fading out the task view when it's getting dismissed.

Test: manual
Fixes: 186531762
Fixes: 189265196

Change-Id: Ieae67fb89b60136fd81ceb8abde1a0cab8718b27