NIU Actions: Add strings for assistant error dialogs and feature tooltip
The assistant error strings will be used in dialogs that appear when the user clicks the NIU Actions buttons without an appropriate default assistant set on the device. The tooltips will appear to highlight and explain the feature for first-time users. Adding the strings before the dialogs themselves in order to unblock translation efforts.

Bug: 192505939
Bug: 186006700
Test: Manual (code builds, existing strings unchanged on device)
Change-Id: I2c37b5871545e99a6bba8d060c927cbd29ef5081