Fix work profile disable button falsing
Bug: 191996677

Test: before this CL: when tapping on immediately above keyboard when
keyboard is translating or when onApplyWindowInset is called, the
hidden (GONE) view handles touch and triggers work profile to be disabled.

After the CL, the button view is not attached to the parent container
when on personal tab. Also, the onApplyWindowInsets method translations
can only be called when work tab is selected.

Change-Id: Ib3e9a5c4b3280052f9ccb62a199ed256ae4e094f