Do not set user active hints when Shell Transition is on.
Previously, we hint at user active in order to let the launcher know at
an earlier time that the user is about to go home, to provide PiP a
shelf offset. However with Shell Transitions, this is now being done
instantly as the launcher activity is resumed right as user touches down
on the gesture pill. We then no longer need this for shell transitions.

Bug: 165793917
Test: Enable Shell Transitions, go home to enter PIP, and open another
app - see that PIP auto moves downward now that shelf height has changed

Change-Id: I347c885425ce820a42ebfcc15082862d04aa1a7d