Move/Launch task on the TaskView's display.
Add ActivityOptions#setLaunchDisplayId to the display that the TaskView
is currently on. So the task will be moved to the display as the user is
interactiving with.

Bug: 183461428
Test: Manually tested using Exo. Open appliction(e.g.Calculator) on
Exo virtual display, and tapping on the phone's recents to bring the
application back to phone.
The demo video  https://drive.google.com/file/d/1JMncvsI-VPosK4ql2xeVA8csVAF7Rb16/view?usp=sharing&resourcekey=0-PYRw1k3KdtY0-KZNo6OqjQ

Change-Id: Iaed5aed4d698b71b0d0fc2b78843f4aedd586ecb