Show distinct TaskMenus for each task in GroupedTaskView
* For now we only show the App Info option
for each task icon clicked for GroupedTaskView
* Have TaskMenuView operate on a specific task instead
of only a TaskView
* NOTE: getItemInfo() in TaskIdAttributeContainer needs to
dynamically call getItemInfo(Task) because at the time of
creation of the container in TaskView#bind(), the task
object provided initially is a palceholder task which only
has a taskId and no other attributes set.
getItemInfo() needs a non-null baseIntent, which the
placeholder task doesn't have.
Right fix for this is to have GestureState hold onto
multiple running tasks when in split screen that are
provided by ActivityManagerWrapper when the gesture starts
(but that change is extensive and out of scope w/ this
workaround available)

Bug: 181704764
Test: Open TaskMenuView's w/ either task icon for
GroupedTaskView, opens up corresponding tasks menu.

Change-Id: I70b7c13394ad4980cabbd611cb928bb03d8a2924