Fix the fallback logic for shortcut icons when icons are missing
The icons of shortcuts in general are not persisted in backup payload
and thus are not available on the new device, apps are reposponsible for
republishing those shortcuts on the new device, with the only exception
being pinned shortcuts, whose icons are stored in launcher.db as a
fallback when icons are not avaialble in the shortcut themselves.

A previous CL unexpected excluded deep shortcuts from loading their
fallback icons in launcher.db, causes pinned shortcut to fallback
further to default droid icons, this CL restores the original behavior.

Bug: 202346722
Test: manual
Change-Id: Ie09becd3a0dc1ca1e7c2573fe1b68b61a58c2091