Tapl Widgets: Don't use container width as gesture margin
The test used margin that was equal to the width of the container when
injecting a backward swipe. This means the swipe was injected at an x
value of 0 in the View's local coordinates. When injecting such events
into the system, the coordinates will undergo several transformations,
at which point floating point precision errors could accumulate. This
could result in the event being dispatched outside the boundary of the
view.

Change the injection margin so that the swipe is injected in the middle
of the View instead of at its edge.

Bug: 201777251
Bug: 207146693
Test: atest TaplTestsLauncher3
Change-Id: I4082c2845aef033a6fbe41070061866d6048e21d