Remove keepWindowOpaque and have the window always fade out.
We built this method so that the window would have
something to fade out on top of. But with longer icon loading times,
this causes a buggy experience since the window is opaque
for the entire duration of the animation.

Having the window always fade out is a better UX for now.
Next step is to always init the FloatingIconView with the
in-memory icon, and then possibly caching both layers
of AdaptiveIcon.

Test: restart phone, swipe up to home
      restart phone, back swipe to home (This flow never called keepWindowOpaque)
Bug: 207389002

Change-Id: Ida64829c60881786ed91e8641cdf7a278b37e201