Always draw background for task thumbnail
Always draw background for task thumbnail because the task snapshot
might be translucent or partially empty.

For example, when drag-to-dismiss split screen, it will reparent tasks
from dismissing split root to the bottom of display. Task thumbnail of
those tasks might be partially empty because the snapshot was taken in
new bounds inherits from display but with split screen content.

Fix: 200096421
Test: drag to dismiss split, observed task thumbnail no longger been
      cropped.
Change-Id: I5d04daa55a558f79ccf5127080088bd0698d06d8