Set launch display id to an action performed on BaseQuickStepLauncher
Without this change, tapping all apps of work profile wouldn't bring
the work profile app back to phone.
Add ActivityOptions#setLaunchDisplayId to the display that the given
view is currently on. So the activity will be launched to the display
as the user is interacting with.

This change also modified BaseDraggingActivity which is the parent
class of BaseQuickstepLauncher. And the RecentsActivity is extending
BaseDraggingActivity.

Bug: 204164763
Test: Manually tested using Exo. Open work appliction(e.g.Calendar)
on Exo virtual display, and tapping on the phone's Calendar on work
profile to bring the application back to phone.
Demo video: https://drive.google.com/file/d/18LWMyDz5kYj8IwAwu5EOsBk3PsjEx1zf/view?usp=sharing&resourcekey=0-7eGtqiWZDciSCfFce29MtA

Change-Id: I31d822f6df7222939413f13da850230f46ba8001