Launch activity if click on a TaskView which has the task running on different display wit root.
If click on the TaskView which has the task running on differnt display
with root, it means we are trying to pull the task to the root display.
So we should also launch the activity to the right display even it is in
live tile mode.
Also add null check for getRootViewDisplayId().

Bug: 207085426
Bug: 210075742
Test: manual test with app streaming wit phonehub notification
Change-Id: Icdaa6dfa26345d271f3505a0226cf434696f40da