Top row of first workspace screen occupied only when QSB is placed at the top.
This CL facilitates making the Feature Flag QSB_ON_FIRST_SCREEN as false
for LauncherLilyGoogle to make sure that that the top row on the first
screen is not blocked for QSB. Hinging it on BuildConfig allows it be
overwritten in LauncherLilyGoogle.

Bug: 208136186
Bug: 188511548

Test: Manually tested that top row of the first screen is not occupied
and does not clear the shortcuts after a power cycle for LauncherLilyGoogle.

Change-Id: I9113e2e2f2c0a160a2e1caf0ef3c5f86c6ec54b8