Close open folders when pausing launcher.
Previously, folders wouldn't close after launching an app and quickly navigating back the launcher.
This is because the onStop state which contains a call to AbstractFloatingView.closeViews() wasn't being called.
This change adds an additional call to specifically close any open folders when the Launcher enters a paused state.

Bug: 227142635
Test: Verified that the issue is fixed after code change.
I logged the Launcher's lifecycle state to ensure my manual testing was reliable.

Change-Id: I838f992f83afdd754672ebbe85d0eb2a7e894417