Finish recents animation always
Previously, some launcher-side state would be finished, but the
actual animation would never be finished until the next touch-down.

This is a problem for shell-transitions because it relies on
animations finishing. This change handles the scenario where
the gesture finishes before the animation even starts by slotting
in a placeholder listener that just immediately finishes. If the
animation has started but there was no motion, then it directly
finishes the recents animation.

Bug: 228898164
Test: open an app, tap on the bottom-bar. With logging it should
      be evident that the transition finishes.
Change-Id: I27e00ceb4b21a2c37dc556bfecb4a90ef7155a85